Move CLI implementation to the generator package.
